prepare frontend
The seaorm generated entities are not useable in a wasm target just like
that since they depend on crates that do not compile for wasm. So I
copied the entities just for the struct definitions and deleted the
excess stuff.
Maybe generating just the structs should be added as a flag for the
seaorm code generator.